Rename phpunit/modules/core/modules.php to make other tests availlable
