+++ tests/phpunit/phpunit.xml
@@ -128,7 +128,9 @@
       <file>./modules/core/functions.php</file>
       <file>./modules/core/message_functions.php</file>
       <file>./modules/core/message_list_functions.php</file>
-      <file>./modules/core/modules.php</file>
+      <file>./modules/core/handler_modules.php</file>
+      <file>./modules/core/output_modules.php</file>
+      <file>./modules/core/output_modules_debug.php</file>
     </testsuite>
   </testsuites>
   <source>
